+++ CHANGELOG.md
@@ -26,6 +26,11 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://bitbucket.org/atlassian/jira-actions/branches/compare/master%0Drelease-3.2.0
 
+### Added
+- Provide Jira Core scenario. Resolves [JPERF-373].
+
+[JPERF-373]: https://ecosystem.atlassian.net/browse/JPERF-373
+
 ## [3.2.0] - 2019-01-08
 [3.2.0]: https://bitbucket.org/atlassian/jira-actions/branches/compare/release-3.2.0%0Drelease-3.1.0